examples/chatbots/python/requirements.txt
strands-agents==1.15.0
mcp==1.20.0
uvicorn==0.38.0
boto3==1.40.67

# For testing, this module is installed from local files.
# Uncomment this line to build using the module from PyPi
# run-mcp-servers-with-aws-lambda >= 0.0.1
